05a._Expose_Integrate_Webhook_System/python-webhook/pyproject.toml
[tool.poetry]
name = "python-webhook"
version = "0.1.0"
description = ""
authors = ["Roed <Geofery10@hotmail.com>"]
readme = "README.md"
packages = [{include = "python_webhook"}]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.28.0"
fastapi = "^0.110.0"
mysql-connector-python = "^8.3.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
